+++ R/nesting.R
@@ -6,6 +6,8 @@
 #' @param param The nested parameter to add data to.
 #' @param .serie Serie's index to add the data to, if
 #' `NULL` then it is added to all.
+#' @param .data A dataset to use, if none are specified than
+#' the original dataset passed to `e_charts` is used.
 #'
 #' @details For instance, \code{\link{e_funnel}} lets you pass \code{values} and \code{labels}
 #' (from your initial data.frame) which corresponds to \code{name} and \code{value} in the
@@ -55,13 +57,18 @@
 #'     fontStyle
 #'   )
 #' @export
-e_add <- function(e, param, ..., .serie = NULL) {
+e_add <- function(e, param, ..., .serie = NULL, .data = NULL) {
   if (missing(e) || missing(param)) {
     stop("missing e or param", call. = FALSE)
   }
 
-  for (i in seq_along(e$x$data)) {
-    data <- e$x$data[[i]] |>
+  if(is.null(.data))
+    ds <- e$x$data
+  else
+    ds <- map_grps_(.data, e$x$tl)
+
+  for (i in seq_along(ds)) {
+    data <- ds[[i]] |>
       dplyr::select(...)
 
     data <- apply(data, 1, as.list)