+++ NEWS.md
@@ -4,6 +4,7 @@
 [#347](https://github.com/JohnCoene/echarts4r/issues/347)
 - Added `.serie` argument to `e_add` to allow adding data to
 a specific serie.
+- Added `.data` argument to `e_add`.
 
 # echarts4r 0.4.1